pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "fastapi-cli>=0.0.5",
    "sqlalchemy>=2.0.36",
    "typer>=0.14.0",
    "uvicorn>=0.32.1",
    "pydantic-settings>=2.6.1",
    "aiosqlite>=0.20.0",
    "bcrypt>=4.2.1",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "strawberry-graphql>=0.253.1",
    "pydantic[email]>=2.10.2",
    "python-multipart>=0.0.19",
]
[tool.pyright]
venvPath = "."
venv = ".venv"
